post response 수정

+++ src/main/java/clone/colley/service/PostService.java
@@ -25,22 +25,34 @@ public class PostService {
 
     //게시글 수정
     @Transactional
-    public void postUpdate(PostRequestDto requestDto, Long postId) {
+    public Boolean postUpdate(PostRequestDto requestDto, Long postId,UserDetailsImpl userDetails) {
         Posts posts=postRepository.findById(postId).orElseThrow(
                 ()->new NullPointerException("수정할 게시글이 없습니다.")
         );
-        posts.setTitle(requestDto.getTitle());
-        posts.setContent(requestDto.getContent());
-        posts.setImgUrl(requestDto.getImgUrl());
+        if(posts.getUser().equals(userDetails.getUser())){
+            posts.setTitle(requestDto.getTitle());
+            posts.setContent(requestDto.getContent());
+            posts.setImgUrl(requestDto.getImgUrl());
+            return true;
+        } else{
+            return false;
+        }
+
     }
 
     //게시글 삭제
     @Transactional
-    public void deletePost(Long postId) {
+    public Boolean deletePost(Long postId,UserDetailsImpl userDetails) {
         Posts posts=postRepository.findById(postId).orElseThrow(
                 ()->new NullPointerException("삭제할 게시글이 없습니다.")
         );
-        postRepository.deleteById(posts.getPostId());
+        if(posts.getUser().equals(userDetails.getUser())){
+            postRepository.deleteById(posts.getPostId());
+            return true;
+        } else{
+            return false;
+        }
+
     }
 
     //게시글 상세조회
@@ -57,6 +69,7 @@ public class PostService {
         responseDto.setLikeCnt(posts.getLikeCnt());
         responseDto.setCommentCnt(posts.getCommentCnt());
         responseDto.setImaUrl(posts.getImgUrl());
+        responseDto.setUsername(posts.getUser().getUsername());
         responseDto.setProfileUrl(posts.getUser().getProfileUrl());
         responseDto.setNickname(posts.getUser().getNickname());
         return  responseDto;
